### 2025-01-26 #7722 @editorialbot
[PRE REVIEW]: Million Points of Light (MPoL): a PyTorch library for radio interferometric imaging and inference
**Submitting author:** @iancze (<a href="http://orcid.org/0000-0002-1483-8811">Ian Czekala</a>)
**Repository:** https://github.com/MPoL-dev/MPoL
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0-alpha
