### 2025-01-26 #7722 @warrickball
Hi @iancze, thanks for your submission to JOSS. While co-publication in other journals doesn't necessarily rule out a publication in JOSS, the [2023 PASP article](https://ui.adsabs.harvard.edu/abs/2023PASP..135f4503Z) does make repeated and specific reference to MPoL, so I'm going to ask the editorial board to consider whether the implementation of the software itself reflects a sufficient substantial scientific effort for JOSS's purposes. This review should take a week or two. Sorry for the additional wait!
