### 2025-02-04 #7722 @warrickball
@iancze — There was enough support on the editorial board that I'm happy to continue with this submission. I'll note also that it looks like about 1600 lines of code have been added to the package since the PASP paper was published.

That said, all the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your continuing patience!
