### 2025-07-14 MPoL-dev/MPoL#263 @mkolopanis
Minimum python verison
The pyproject.toml requires define python3.8 as the minimum working python.

https://github.com/MPoL-dev/MPoL/blob/f1facdb3737a2b45c5308f0ebd9ee60583edf0f1/pyproject.toml#L26

However, continuous integration is only run on 3.10-3.12 and when attempting run this package in python 3.8 yields typing related errors:

```bash
-> % python -c "from mpol import utils"
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "/home/matthew/src/MPoL/.venv-3.8/lib/python3.8/site-packages/mpol/utils.py", line 135, in <module>
    ) -> npt.NDArray[np.floating[Any]]:
TypeError: Type subscription requires python >= 3.9
```
There are ways to back port some typing like this to 3.8 but I am curious if you desire a minimum python of 3.10 based on the repository's automated testing.
