### 2025-07-16 MPoL-dev/MPoL#263 @iancze
The minimum python version technically should have been 3.10 ([now 3.11](https://scientific-python.org/specs/spec-0000/)) but the reason we didn't set this is as a hard requirement in pyproject.toml was that for a long time [modular CASA](https://casadocs.readthedocs.io/en/stable/notebooks/introduction.html#Compatibility) only worked with python 3.8, and some users would want to have both MPoL and CASA installed into the same environment.

It seems like modular CASA is now at a state that bumping MPoL to a 3.10 minimum version of Python is acceptable.
