### 2025-07-16 MPoL-dev/MPoL#263 @mkolopanis
I think this is hitting on the crux of my particular issue too. As I am attempting to move through the examples for the review, I am finding that I require extra python environments to support the casa requirements because casa's python tools tend to have a longer delay for updating python versions.
