### 2025-08-04 MPoL-dev/MPoL#264 @mkolopanis
Coordinates Docs not rendering properly
I believe the coordinates API documentation is having a slip-up in the doc auto-gen. I see the image below on firefox, but all other doc pages render nicely with ivars and math blocks.

`check_data_fit` function seem to render well, maybe it is just a tabbing issue in the parameters for coordinates?
https://github.com/MPoL-dev/MPoL/blob/f1facdb3737a2b45c5308f0ebd9ee60583edf0f1/src/mpol/coordinates.py#L21-L26
