### 2025-08-05 #8419 @plaplant
@mkolopanis @kartographer thanks very much for going through the reviewer checklists! It seems like most of the items have been ticked off, but there are still a few that might require changes to address.

@mkolopanis I see you've made a coupe of issues on the repo already. If there are additional points to address, please feel free to open additional issues or respond to this thread.

@kartographer if you notice changes that should be made, please feel free to open issues on the [main code repository](https://github.com/MPoL-dev/MPoL) or respond to this thread. If not, then please respond to this issue so we know that it's acceptable as is.

Thanks again for reviewing! Let me know if you have any questions.
