### 2025-08-13 24c0bbc
Update paper.md authors with Jane Huang, finally!
